+++ packages/styled-components/components/container/index.js
@@ -37,9 +37,13 @@ const Container = (props) => {
   }
 
   // Setup a max-width style.
-  if ('string' === typeof maxWidth && undefined !== widths[maxWidth]) {
+  if (
+    'string' === typeof maxWidth &&
+    undefined !== widths[maxWidth] &&
+    0 !== maxWidth.length
+  ) {
     style.maxWidth = `${widths[maxWidth]}px`; // Use one of our predefined sizes.
-  } else if ('string' === typeof maxWidth) {
+  } else if ('string' === typeof maxWidth && 0 !== maxWidth.length) {
     style.maxWidth = maxWidth; // Use the string value 1:1.
   } else if ('number' === typeof maxWidth) {
     style.maxWidth = `${maxWidth}px`; // Use the interger value as pixels.