+++ packages/styled/hooks/useStandardProps.js
@@ -22,18 +22,32 @@ export const replaceWithSiteTheme = (styleObject, siteTheme) => (
        *   Match:
        *    {color: theme.colors.black} => {color: #000;}
        *
+       *   Recursive Match:
+       *    {color: theme.fonts.color} =>
+       *      {color: theme.colors.black} =>
+       *        {color: #000}
+       *
        *   No Match:
        *     {color: #000} => {color: #000;}
        */
-      const themeStyle = get(
-        siteTheme,
-        styleObject[property],
-        styleObject[property]
-      );
+      let returnValue = styleObject[property];
+      let defaultValue = returnValue;
+
+      // Recursively look for the returned value in the theme provider until
+      // the default is returned.
+      do {
+        defaultValue = returnValue;
+
+        returnValue = get(
+          siteTheme,
+          returnValue,
+          defaultValue
+        );
+      } while (returnValue !== defaultValue);
 
       return {
         ...acc,
-        [property]: themeStyle,
+        [property]: returnValue,
       };
     }, styleObject)
 );